[RENDER] Adding struct aligment requirements: * 16 byte alignment in the CPU buffer, explicit (using the newly introduced PEGASUS_ALIGN macros) * std140 layout explicit in the uniform block in the shader * manual padding of shader uniform blocks: some drivers (like ATI) don't pad the shader uniform block with bytes, which is a requirement very vague in the std140 standard * Changes also include removing certain operations that could affect endianness in some machines and crash in a draw call. The particular operation being caching the usage flags in a 64 bit variable.
Additionally:
* Fixed RenderContext to not swap device context if one currently exists. Also put all the device context (and render context) in the thread local storage, to force a crash if the user calls a render function from the wrong thread.

Jobs: job000126
[git-p4: depot-paths = "//Pegasus/DL/": change = 1135]

+++ include/Pegasus/Render/Render.h
@@ -51,7 +51,7 @@ namespace Render
     struct Buffer
     {
         int mSize;
-        unsigned long long mInternalData; //! 64 bit compressed data, used internally
+        int mInternalData; //! 64 bit compressed data, used internally
     public:
         Buffer()
            : mInternalData(0), mSize(0)
